# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SMS Hub API",
        "description": "SMS Hub API Documentation - A comprehensive SMS gateway service with multiple provider support",
        "contact": {
            "name": "SMS Hub Support",
            "email": "support@smshub.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Development Server"
        }
    ],
    "paths": {
        "/api/v1/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get list of SMS messages",
                "description": "Retrieve a paginated list of SMS messages for the authenticated user",
                "operationId": "getMessages",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by message status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "sent",
                                "delivered",
                                "failed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Message"
                                            }
                                        },
                                        "links": {
                                            "$ref": "#/components/schemas/PaginationLinks"
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Send SMS message",
                "description": "Send a new SMS message through the SMS Hub",
                "operationId": "sendMessage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "to",
                                    "message"
                                ],
                                "properties": {
                                    "to": {
                                        "description": "Recipient phone number",
                                        "type": "string",
                                        "example": "+998901234567"
                                    },
                                    "message": {
                                        "description": "SMS message content",
                                        "type": "string",
                                        "example": "Hello, this is a test message!"
                                    },
                                    "provider": {
                                        "description": "SMS provider to use",
                                        "type": "string",
                                        "example": "eskiz"
                                    },
                                    "sender_id": {
                                        "description": "Sender ID",
                                        "type": "string",
                                        "example": "SMSHub"
                                    },
                                    "callback_url": {
                                        "description": "Webhook URL for delivery reports",
                                        "type": "string",
                                        "example": "https://example.com/webhook"
                                    },
                                    "priority": {
                                        "description": "Message priority (1-5)",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message sent successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Message"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Message sent successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/messages/{id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Get message details",
                "description": "Retrieve detailed information about a specific SMS message",
                "operationId": "getMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Message ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Message not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Messages"
                ],
                "summary": "Update message",
                "description": "Update an existing SMS message (only if not sent)",
                "operationId": "updateMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Message ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "message": {
                                        "description": "Updated SMS message content",
                                        "type": "string"
                                    },
                                    "priority": {
                                        "description": "Updated message priority (1-5)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Message updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Message"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Message updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - message already sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Message not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Delete message",
                "description": "Delete an SMS message (only if not sent)",
                "operationId": "deleteMessage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Message ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Message deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - message already sent",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Message not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Message": {
                "title": "SMS Message",
                "description": "SMS message object",
                "properties": {
                    "id": {
                        "description": "Message ID",
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "to": {
                        "description": "Recipient phone number",
                        "type": "string",
                        "example": "+998901234567"
                    },
                    "message": {
                        "description": "SMS message content",
                        "type": "string",
                        "example": "Hello, this is a test message!"
                    },
                    "provider": {
                        "description": "SMS provider used",
                        "type": "string",
                        "example": "eskiz"
                    },
                    "sender_id": {
                        "description": "Sender ID",
                        "type": "string",
                        "example": "SMSHub"
                    },
                    "status": {
                        "description": "Message status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "sent",
                            "delivered",
                            "failed"
                        ],
                        "example": "sent"
                    },
                    "priority": {
                        "description": "Message priority (1-5)",
                        "type": "integer",
                        "example": 1
                    },
                    "cost": {
                        "description": "Message cost",
                        "type": "number",
                        "format": "float",
                        "example": 0.05
                    },
                    "provider_message_id": {
                        "description": "Provider's message ID",
                        "type": "string",
                        "example": "provider_12345"
                    },
                    "callback_url": {
                        "description": "Webhook URL for delivery reports",
                        "type": "string",
                        "example": "https://example.com/webhook"
                    },
                    "sent_at": {
                        "description": "Sent timestamp",
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "delivered_at": {
                        "description": "Delivery timestamp",
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-15T10:35:00Z"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-15T10:30:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "datetime",
                        "example": "2024-01-15T10:35:00Z"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "title": "Error Response",
                "description": "Standard error response",
                "properties": {
                    "error": {
                        "description": "Error type",
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "message": {
                        "description": "Error message",
                        "type": "string",
                        "example": "Invalid or missing authentication token"
                    },
                    "code": {
                        "description": "HTTP status code",
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
...
            "PaginationLinks": {
                "title": "Pagination Links",
                "description": "Pagination links",
                "properties": {
                    "first": {
                        "description": "First page URL",
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/messages?page=1"
                    },
                    "last": {
                        "description": "Last page URL",
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/messages?page=10"
                    },
                    "prev": {
                        "description": "Previous page URL",
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/messages?page=2",
                        "nullable": true
                    },
                    "next": {
                        "description": "Next page URL",
                        "type": "string",
                        "example": "http://localhost:8000/api/v1/messages?page=4",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "title": "Pagination Meta",
                "description": "Pagination metadata",
                "properties": {
                    "current_page": {
                        "description": "Current page number",
                        "type": "integer",
                        "example": 3
                    },
                    "from": {
                        "description": "First item number on current page",
                        "type": "integer",
                        "example": 31
                    },
                    "last_page": {
                        "description": "Last page number",
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "description": "Items per page",
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "description": "Last item number on current page",
                        "type": "integer",
                        "example": 45
                    },
                    "total": {
                        "description": "Total number of items",
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "scopes": {
                            "read": "Read access",
                            "write": "Write access",
                            "admin": "Admin access"
                        }
                    },
                    "password": {
                        "tokenUrl": "http://localhost:8000/oauth/token",
                        "scopes": {
                            "read": "Read access",
                            "write": "Write access",
                            "admin": "Admin access"
                        }
                    }
                }
            },
            "bearer": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearer": []
        }
    ]
}